Streamlining
Creates new file Grade Update Checker.py to declutter the main file from
a lot of the parsing. Selenium now is initialized out of the monitor to
make it seem like it's invisible. new file Datareader.py writes the
parsed data into an Excel file with different charts and tables
representing analysis of the data. This is the final iteration of this
project that invovles Selenium. Currently it takes 30 seconds maximum to
download and write data out, which is much longer than it would take to
just login to the site and browse.